+++ engines/default/coll_set.c
@@ -222,7 +222,7 @@ static set_elem_item *do_set_elem_alloc(const uint32_t nbytes, const void *cooki
         assert(elem->slabs_clsid > 0);
 
         elem->refcount    = 0;
-        elem->nbytes      = nbytes;
+        elem->nbytes      = (uint16_t)nbytes;
         elem->status = ELEM_STATUS_UNLINKED; /* unlinked state */
     }
     return elem;
